x/sequencer/types/message_make_block_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"ibc_sequencer/testutil/sample"
)

func TestMsgMakeBlock_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgMakeBlock
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgMakeBlock{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgMakeBlock{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
